+++ scripts/mflt-build-id/tasks/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) Memfault, Inc.
+# See License.txt for details
+#
+
+from mflt.testing.task_templates import make_test_task
+
+test = make_test_task()